+++ iivs/dhm/data/phase/__init__.py
@@ -4,6 +4,7 @@ __all__ = (
     "PhaseBinFolder",
     "PhaseBinHeader",
     "PhaseBinList",
+    "PhaseBounds",
     "PhaseFloatSequence",
     "PhaseImageSequence",
     "PhaseNpyFolder",
@@ -18,7 +19,9 @@ __all__ = (
     "load_phase_txt",
     "read_phase_bin_header",
     "read_phase_txt_header",
+    "read_phbounds",
     "save_phase_bin",
+    "write_phbounds",
 )
 
 from iivs.dhm.data.phase.base import (
@@ -34,6 +37,7 @@ from iivs.dhm.data.phase.bin import (
     read_phase_bin_header,
     save_phase_bin,
 )
+from iivs.dhm.data.phase.bounds import PhaseBounds, read_phbounds, write_phbounds
 from iivs.dhm.data.phase.core import PhaseUnit, convert_phase_unit
 from iivs.dhm.data.phase.npy import PhaseNpyFolder
 from iivs.dhm.data.phase.tif import PhaseTifFolder, PhaseTifList